+++ lib/store/helpers/remove/index.ts
@@ -1,0 +1,15 @@
+import { removeChoiceFromRules } from "./removeChoiceFromRules";
+import { removeChoiceFromNextActions } from "./removeChoiceFromNextActions";
+import { remove } from "./remove";
+import { removeSubItem } from "./removeSubItem";
+import { removeChoice } from "./removeChoice";
+import { removeSubChoice } from "./removeSubChoice";
+
+export {
+  removeChoiceFromRules,
+  removeChoiceFromNextActions,
+  remove,
+  removeSubItem,
+  removeChoice,
+  removeSubChoice,
+};
