+++ lib/store/useTemplateStore.tsx
@@ -7,12 +7,10 @@ import React, { createContext, useRef, useContext, useEffect } from "react";
 import { createStore } from "zustand";
 import { useStoreWithEqualityFn } from "zustand/traditional";
 import { immer } from "zustand/middleware/immer";
-import { original } from "immer";
 import { shallow } from "zustand/shallow";
 import { persist, subscribeWithSelector } from "zustand/middleware";
 import update from "lodash.set";
 import unset from "lodash.unset";
-import { getParentIndex } from "@lib/utils/form-builder/getPath";
 
 /**
  * Internal dependencies
@@ -22,14 +20,7 @@ import { getPathString } from "../utils/form-builder/getPath";
 import { TreeRefProvider } from "@formBuilder/components/shared/right-panel/treeview/provider/TreeRefProvider";
 import { FlowRefProvider } from "@formBuilder/[id]/edit/logic/components/flow/provider/FlowRefProvider";
 import { initializeGroups } from "@formBuilder/components/shared/right-panel/treeview/util/initializeGroups";
-import {
-  removeElementById,
-  removeById,
-  getSchemaFromState,
-  cleanInput,
-  removeGroupElement,
-} from "../utils/form-builder";
-import { decrementChoiceIds, decrementNextActionChoiceIds } from "@lib/formContext";
+import { getSchemaFromState, cleanInput } from "../utils/form-builder";
 import { Language } from "../types/form-builder-types";
 import { defaultForm } from "./defaults";
 import { storageOptions } from "./storage";
@@ -45,6 +36,16 @@ import {
   addSubChoice,
   duplicateElement,
 } from "./helpers/add";
+
+import {
+  removeChoiceFromRules,
+  removeChoiceFromNextActions,
+  remove,
+  removeSubItem,
+  removeChoice,
+  removeSubChoice,
+} from "./helpers/remove";
+
 import { moveUp, moveDown, subMoveUp, subMoveDown } from "./helpers/move";
 
 const createTemplateStore = (initProps?: Partial<InitialTemplateStoreProps>) => {
@@ -130,67 +131,12 @@ const createTemplateStore = (initProps?: Partial<InitialTemplateStoreProps>) => 
             addChoice: addChoice(set),
             addLabeledChoice: addLabeledChoice(set),
             addSubChoice: addSubChoice(set),
-            removeChoiceFromRules: (elId: string, choiceIndex: number) => {
-              set((state) => {
-                const choiceId = `${elId}.${choiceIndex}`;
-                const rules = decrementChoiceIds({ formElements: state.form.elements, choiceId });
-                state.form.elements.forEach((element) => {
-                  // If element id is in the rules array, update the conditionalRules property
-                  if (rules[element.id]) {
-                    element.properties.conditionalRules = rules[element.id];
-                  }
-                });
-              });
-            },
-            removeChoiceFromNextActions: (elId: string, choiceIndex: number) => {
-              set((state) => {
-                const choiceId = `${elId}.${choiceIndex}`;
-                const groups = decrementNextActionChoiceIds({ ...state.form.groups }, choiceId);
-                state.form.groups = groups;
-              });
-            },
-            remove: (elementId, groupId = "") => {
-              set((state) => {
-                const allowGroups = state.allowGroupsFlag;
-                state.form.elements = removeElementById(state.form.elements, elementId);
-                state.form.layout = removeById(state.form.layout, elementId);
-
-                if (allowGroups && groupId && state.form.groups) {
-                  const groups = removeGroupElement(
-                    { ...original(state.form.groups) },
-                    groupId,
-                    elementId
-                  );
-                  state.form.groups = { ...groups };
-                }
-              });
-            },
-            removeSubItem: (elId, elementId) =>
-              set((state) => {
-                const parentIndex = getParentIndex(elId, state.form.elements);
-
-                if (parentIndex === undefined) return;
-
-                const subElements = state.form.elements[parentIndex].properties?.subElements;
-                if (subElements) {
-                  state.form.elements[parentIndex].properties.subElements = removeElementById(
-                    subElements,
-                    elementId
-                  );
-                }
-              }),
-            removeChoice: (elIndex, choiceIndex) =>
-              set((state) => {
-                state.form.elements[elIndex].properties.choices?.splice(choiceIndex, 1);
-              }),
-            removeSubChoice: (elId, subIndex, choiceIndex) =>
-              set((state) => {
-                const parentIndex = getParentIndex(elId, state.form.elements);
-                if (parentIndex === undefined) return;
-                state.form.elements[parentIndex].properties.subElements?.[
-                  subIndex
-                ].properties.choices?.splice(choiceIndex, 1);
-              }),
+            removeChoiceFromRules: removeChoiceFromRules(set),
+            removeChoiceFromNextActions: removeChoiceFromNextActions(set),
+            remove: remove(set),
+            removeSubItem: removeSubItem(set),
+            removeChoice: removeChoice(set),
+            removeSubChoice: removeSubChoice(set),
             getChoice: (elId, choiceIndex) => {
               const elIndex = get().form.elements.findIndex((el) => el.id === elId);
               return get().form.elements[elIndex]?.properties.choices?.[choiceIndex];
